Extract workspace dialog views
