+++ Sources/QuillCodeApp/WorkspaceSwiftUIView.swift
@@ -466,659 +466,6 @@ public struct QuillCodeWorkspaceView: View {
     }
 }
 
-private enum QuillCodeWorktreeSheet: String, Identifiable {
-    case create
-    case remove
-
-    var id: String { rawValue }
-}
-
-private struct QuillCodeThreadRenameDraft: Identifiable, Hashable {
-    var threadID: UUID
-    var title: String
-
-    var id: UUID { threadID }
-}
-
-private struct QuillCodeThreadRenameView: View {
-    var draft: QuillCodeThreadRenameDraft
-    var onCancel: () -> Void
-    var onSave: (UUID, String) -> Void
-
-    @State private var title: String
-
-    init(
-        draft: QuillCodeThreadRenameDraft,
-        onCancel: @escaping () -> Void,
-        onSave: @escaping (UUID, String) -> Void
-    ) {
-        self.draft = draft
-        self.onCancel = onCancel
-        self.onSave = onSave
-        self._title = State(initialValue: draft.title)
-    }
-
-    private var canSave: Bool {
-        !title.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty
-    }
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 16) {
-            Text("Rename Chat")
-                .font(.title2.weight(.semibold))
-            TextField("Chat title", text: $title)
-                .textFieldStyle(.roundedBorder)
-                .onSubmit {
-                    if canSave {
-                        onSave(draft.threadID, title)
-                    }
-                }
-            HStack {
-                Spacer()
-                Button("Cancel", action: onCancel)
-                Button("Save") {
-                    onSave(draft.threadID, title)
-                }
-                .keyboardShortcut(.defaultAction)
-                .disabled(!canSave)
-            }
-        }
-        .padding(22)
-        .frame(width: 380)
-    }
-}
-
-private struct QuillCodeProjectRenameDraft: Identifiable, Hashable {
-    var projectID: UUID
-    var name: String
-
-    var id: UUID { projectID }
-}
-
-private struct QuillCodeProjectRenameView: View {
-    var draft: QuillCodeProjectRenameDraft
-    var onCancel: () -> Void
-    var onSave: (UUID, String) -> Void
-
-    @State private var name: String
-
-    init(
-        draft: QuillCodeProjectRenameDraft,
-        onCancel: @escaping () -> Void,
-        onSave: @escaping (UUID, String) -> Void
-    ) {
-        self.draft = draft
-        self.onCancel = onCancel
-        self.onSave = onSave
-        self._name = State(initialValue: draft.name)
-    }
-
-    private var canSave: Bool {
-        !name.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty
-    }
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 16) {
-            Text("Rename Project")
-                .font(.title2.weight(.semibold))
-            TextField("Project name", text: $name)
-                .textFieldStyle(.roundedBorder)
-                .onSubmit {
-                    if canSave {
-                        onSave(draft.projectID, name)
-                    }
-                }
-            HStack {
-                Spacer()
-                Button("Cancel", action: onCancel)
-                Button("Save") {
-                    onSave(draft.projectID, name)
-                }
-                .keyboardShortcut(.defaultAction)
-                .disabled(!canSave)
-            }
-        }
-        .padding(22)
-        .frame(width: 380)
-    }
-}
-
-private struct QuillCodeCommandPaletteView: View {
-    var commands: [WorkspaceCommandSurface]
-    @Binding var query: String
-    var onSelectCommand: (WorkspaceCommandSurface) -> Void
-    var onClose: () -> Void
-
-    @State private var selectedCommandID: String?
-    @FocusState private var isSearchFocused: Bool
-
-    private var results: [WorkspaceCommandSurface] {
-        WorkspaceCommandPalette.rankedCommands(commands, matching: query)
-    }
-
-    private var groups: [WorkspaceCommandGroupSurface] {
-        WorkspaceCommandPalette.groupedCommands(commands, matching: query)
-    }
-
-    private var enabledResults: [WorkspaceCommandSurface] {
-        results.filter(\.isEnabled)
-    }
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 16) {
-            HStack(alignment: .center, spacing: 12) {
-                VStack(alignment: .leading, spacing: 4) {
-                    Text("Command palette")
-                        .font(.title2.weight(.semibold))
-                    Text("Run actions, or type / to insert slash commands.")
-                        .font(.callout)
-                        .foregroundStyle(QuillCodePalette.muted)
-                }
-                Spacer()
-                Button("Close", action: onClose)
-                    .keyboardShortcut(.cancelAction)
-            }
-
-            HStack(spacing: 10) {
-                TextField("Search commands, > actions, / slash", text: $query)
-                    .textFieldStyle(.roundedBorder)
-                    .focused($isSearchFocused)
-                    .onSubmit(selectHighlightedCommand)
-                if let label = activeScopeLabel {
-                    Text(label)
-                        .font(.caption.weight(.semibold))
-                        .foregroundStyle(QuillCodePalette.blue)
-                        .padding(.horizontal, 10)
-                        .padding(.vertical, 6)
-                        .background(QuillCodePalette.selection)
-                        .clipShape(RoundedRectangle(cornerRadius: 9, style: .continuous))
-                }
-            }
-
-            if results.isEmpty {
-                VStack(spacing: 8) {
-                    Image(systemName: "command")
-                        .font(.title2)
-                        .foregroundStyle(QuillCodePalette.muted)
-                    Text("No matching commands")
-                        .font(.headline)
-                    Text("Try a command name or shortcut.")
-                        .font(.callout)
-                        .foregroundStyle(QuillCodePalette.muted)
-                }
-                .frame(maxWidth: .infinity, minHeight: 180)
-            } else {
-                ScrollView {
-                    LazyVStack(alignment: .leading, spacing: 10) {
-                        ForEach(groups) { group in
-                            VStack(alignment: .leading, spacing: 6) {
-                                Text(group.title)
-                                    .font(.caption.weight(.semibold))
-                                    .foregroundStyle(QuillCodePalette.muted)
-                                    .textCase(.uppercase)
-                                ForEach(group.commands) { command in
-                                    commandButton(command)
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        .padding(20)
-        .frame(width: 560, height: 520)
-        .background(QuillCodePalette.background)
-        .onAppear {
-            ensureSelection()
-            DispatchQueue.main.async {
-                isSearchFocused = true
-            }
-        }
-        .onDisappear {
-            isSearchFocused = false
-        }
-        .onChange(of: query) { _, _ in
-            ensureSelection()
-        }
-        .onMoveCommand { direction in
-            switch direction {
-            case .up:
-                moveSelection(by: -1)
-            case .down:
-                moveSelection(by: 1)
-            default:
-                break
-            }
-        }
-    }
-
-    private func commandButton(_ command: WorkspaceCommandSurface) -> some View {
-        Button {
-            selectedCommandID = command.id
-            onSelectCommand(command)
-        } label: {
-            HStack(spacing: 12) {
-                Image(systemName: systemImage(for: command.id))
-                    .foregroundStyle(command.isEnabled ? QuillCodePalette.blue : QuillCodePalette.muted)
-                    .frame(width: 22)
-                VStack(alignment: .leading, spacing: 3) {
-                    Text(command.title)
-                        .font(.callout.weight(.semibold))
-                    if !command.keywords.isEmpty {
-                        Text(command.keywords.prefix(3).joined(separator: " - "))
-                            .font(.caption)
-                            .foregroundStyle(QuillCodePalette.muted)
-                            .lineLimit(1)
-                    }
-                }
-                Spacer()
-                if let shortcut = command.shortcut {
-                    Text(shortcut)
-                        .font(.caption.monospaced())
-                        .foregroundStyle(QuillCodePalette.muted)
-                }
-            }
-            .padding(12)
-            .background(command.id == selectedCommandID ? QuillCodePalette.selection : QuillCodePalette.panel)
-            .overlay(
-                RoundedRectangle(cornerRadius: 12, style: .continuous)
-                    .stroke(command.id == selectedCommandID ? QuillCodePalette.blue.opacity(0.6) : Color.clear)
-            )
-            .clipShape(RoundedRectangle(cornerRadius: 12, style: .continuous))
-        }
-        .buttonStyle(.plain)
-        .disabled(!command.isEnabled)
-        .help(command.keywords.last ?? command.title)
-    }
-
-    private var activeScopeLabel: String? {
-        let trimmed = query.trimmingCharacters(in: .whitespacesAndNewlines)
-        if trimmed.hasPrefix("/") {
-            return "Slash"
-        }
-        if trimmed.hasPrefix(">") {
-            return "Actions"
-        }
-        return nil
-    }
-
-    private func ensureSelection() {
-        if let selectedCommandID, enabledResults.contains(where: { $0.id == selectedCommandID }) {
-            return
-        }
-        selectedCommandID = enabledResults.first?.id
-    }
-
-    private func moveSelection(by delta: Int) {
-        guard !enabledResults.isEmpty else {
-            selectedCommandID = nil
-            return
-        }
-        let currentIndex = selectedCommandID.flatMap { id in
-            enabledResults.firstIndex(where: { $0.id == id })
-        } ?? 0
-        let nextIndex = (currentIndex + delta + enabledResults.count) % enabledResults.count
-        selectedCommandID = enabledResults[nextIndex].id
-    }
-
-    private func selectHighlightedCommand() {
-        guard let command = enabledResults.first(where: { $0.id == selectedCommandID }) ?? enabledResults.first else {
-            return
-        }
-        onSelectCommand(command)
-    }
-
-    private func systemImage(for commandID: String) -> String {
-        switch commandID {
-        case _ where commandID.hasPrefix(SlashCommandCatalog.commandPaletteIDPrefix):
-            return "slash.circle"
-        case "new-chat":
-            return "square.and.pencil"
-        case "search":
-            return "magnifyingglass"
-        case "find-in-chat":
-            return "text.magnifyingglass"
-        case "add-project":
-            return "folder.badge.plus"
-        case "project-new-chat":
-            return "plus.message"
-        case "project-refresh-context":
-            return "arrow.clockwise"
-        case "project-rename":
-            return "text.cursor"
-        case "project-remove":
-            return "minus.circle"
-        case "toggle-terminal":
-            return "terminal"
-        case "terminal-clear":
-            return "clear"
-        case "toggle-browser":
-            return "globe"
-        case "toggle-activity":
-            return "list.bullet.rectangle"
-        case "toggle-automations":
-            return "clock.arrow.circlepath"
-        case "toggle-memories", "memory-add":
-            return "brain.head.profile"
-        case "toggle-extensions":
-            return "puzzlepiece.extension"
-        case "git-pr-create":
-            return "arrow.up.doc"
-        case "git-pr-checkout":
-            return "arrow.down.doc"
-        case "git-pr-reviewers":
-            return "person.2.badge.gearshape"
-        case "git-pr-labels":
-            return "tag"
-        case "git-pr-merge":
-            return "arrow.triangle.merge"
-        case "git-worktree-list":
-            return "point.3.connected.trianglepath.dotted"
-        case "git-worktree-create":
-            return "plus.rectangle.on.folder"
-        case "git-worktree-remove":
-            return "minus.rectangle"
-        case "settings":
-            return "gearshape"
-        case "keyboard-shortcuts":
-            return "keyboard"
-        case "computer-use-setup":
-            return "display"
-        case "stop-all":
-            return "stop.circle"
-        default:
-            if commandID.hasPrefix("local-env:") {
-                return "hammer"
-            }
-            return "command"
-        }
-    }
-}
-
-private struct QuillCodeKeyboardShortcutsView: View {
-    var commands: [WorkspaceCommandSurface]
-    var onClose: () -> Void
-
-    private var shortcutCommands: [WorkspaceCommandSurface] {
-        commands.filter { $0.shortcut?.isEmpty == false }
-    }
-
-    private var groups: [WorkspaceCommandGroupSurface] {
-        WorkspaceCommandPalette.groupedCommands(shortcutCommands, matching: "")
-    }
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 16) {
-            HStack(alignment: .center, spacing: 12) {
-                VStack(alignment: .leading, spacing: 4) {
-                    Text("Keyboard shortcuts")
-                        .font(.title2.weight(.semibold))
-                    Text("Fast paths for the workspace actions available right now.")
-                        .font(.callout)
-                        .foregroundStyle(QuillCodePalette.muted)
-                }
-                Spacer()
-                Button("Close", action: onClose)
-                    .keyboardShortcut(.cancelAction)
-            }
-
-            ScrollView {
-                LazyVStack(alignment: .leading, spacing: 14) {
-                    ForEach(groups) { group in
-                        VStack(alignment: .leading, spacing: 6) {
-                            Text(group.title)
-                                .font(.caption.weight(.semibold))
-                                .foregroundStyle(QuillCodePalette.muted)
-                                .textCase(.uppercase)
-                            ForEach(group.commands) { command in
-                                HStack(spacing: 12) {
-                                    VStack(alignment: .leading, spacing: 3) {
-                                        Text(command.title)
-                                            .font(.callout.weight(.semibold))
-                                            .foregroundStyle(command.isEnabled ? QuillCodePalette.text : QuillCodePalette.muted)
-                                        if !command.keywords.isEmpty {
-                                            Text(command.keywords.prefix(3).joined(separator: " - "))
-                                                .font(.caption)
-                                                .foregroundStyle(QuillCodePalette.muted)
-                                                .lineLimit(1)
-                                        }
-                                    }
-                                    Spacer()
-                                    Text(command.shortcut ?? "")
-                                        .font(.caption.monospaced().weight(.semibold))
-                                        .foregroundStyle(QuillCodePalette.text)
-                                        .padding(.horizontal, 10)
-                                        .padding(.vertical, 6)
-                                        .background(QuillCodePalette.selection)
-                                        .clipShape(RoundedRectangle(cornerRadius: 8, style: .continuous))
-                                }
-                                .padding(12)
-                                .background(QuillCodePalette.panel)
-                                .overlay(
-                                    RoundedRectangle(cornerRadius: 12, style: .continuous)
-                                        .stroke(Color.white.opacity(0.08))
-                                )
-                                .clipShape(RoundedRectangle(cornerRadius: 12, style: .continuous))
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        .padding(20)
-        .frame(width: 560, height: 520)
-        .background(QuillCodePalette.background)
-    }
-}
-
-private struct QuillCodeSearchView: View {
-    var sidebar: SidebarSurface
-    @Binding var query: String
-    var onSelectThread: (UUID) -> Void
-    var onClose: () -> Void
-
-    @FocusState private var isSearchFocused: Bool
-
-    private var results: [SidebarItemSurface] {
-        sidebar.filteredItems(matching: query)
-    }
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 16) {
-            HStack(alignment: .center, spacing: 12) {
-                VStack(alignment: .leading, spacing: 4) {
-                    Text("Search chats")
-                        .font(.title2.weight(.semibold))
-                    Text("Find a thread by title, model, pinned state, archived state, or transcript text.")
-                        .font(.callout)
-                        .foregroundStyle(QuillCodePalette.muted)
-                }
-                Spacer()
-                Button("Close", action: onClose)
-                    .keyboardShortcut(.cancelAction)
-            }
-
-            TextField("Search chats", text: $query)
-                .textFieldStyle(.roundedBorder)
-                .focused($isSearchFocused)
-                .onSubmit {
-                    if let firstResult = results.first {
-                        onSelectThread(firstResult.id)
-                    }
-                }
-
-            if results.isEmpty {
-                VStack(spacing: 8) {
-                    Image(systemName: "magnifyingglass")
-                        .font(.title2)
-                        .foregroundStyle(QuillCodePalette.muted)
-                    Text("No matching chats")
-                        .font(.headline)
-                    Text("Try a thread title, selected model, pinned, or prior message text.")
-                        .font(.callout)
-                        .foregroundStyle(QuillCodePalette.muted)
-                }
-                .frame(maxWidth: .infinity, minHeight: 180)
-            } else {
-                ScrollView {
-                    LazyVStack(spacing: 8) {
-                        ForEach(results) { item in
-                            Button {
-                                onSelectThread(item.id)
-                            } label: {
-                                HStack(spacing: 12) {
-                                    Image(systemName: item.isPinned ? "pin.fill" : "text.bubble")
-                                        .foregroundStyle(item.isSelected ? QuillCodePalette.blue : QuillCodePalette.muted)
-                                        .frame(width: 22)
-                                    VStack(alignment: .leading, spacing: 4) {
-                                        Text(item.title)
-                                            .font(.callout.weight(.semibold))
-                                            .lineLimit(1)
-                                        Text(item.subtitle + (item.isPinned ? " - pinned" : "") + (item.isArchived ? " - archived" : ""))
-                                            .font(.caption)
-                                            .foregroundStyle(QuillCodePalette.muted)
-                                            .lineLimit(1)
-                                    }
-                                    Spacer()
-                                    if item.isSelected {
-                                        Image(systemName: "checkmark")
-                                            .foregroundStyle(QuillCodePalette.blue)
-                                    }
-                                }
-                                .padding(12)
-                                .background(item.isSelected ? QuillCodePalette.selection : QuillCodePalette.panel)
-                                .clipShape(RoundedRectangle(cornerRadius: 12, style: .continuous))
-                            }
-                            .buttonStyle(.plain)
-                        }
-                    }
-                }
-            }
-        }
-        .padding(20)
-        .frame(width: 560, height: 520)
-        .background(QuillCodePalette.background)
-        .onAppear {
-            DispatchQueue.main.async {
-                isSearchFocused = true
-            }
-        }
-        .onDisappear {
-            isSearchFocused = false
-        }
-    }
-}
-
-private struct QuillCodeWorktreeCreateView: View {
-    @Binding var draft: QuillCodeWorktreeCreateDraft
-    var onCancel: () -> Void
-    var onCreate: () -> Void
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 18) {
-            HStack(alignment: .center) {
-                VStack(alignment: .leading, spacing: 4) {
-                    Text("Create Worktree")
-                        .font(.title2.weight(.semibold))
-                    Text("Create a sibling git worktree for this project.")
-                        .font(.callout)
-                        .foregroundStyle(QuillCodePalette.muted)
-                }
-                Spacer()
-                Image(systemName: "plus.rectangle.on.folder")
-                    .font(.title2)
-                    .foregroundStyle(QuillCodePalette.blue)
-            }
-
-            VStack(alignment: .leading, spacing: 6) {
-                Text("Worktree folder")
-                    .font(.caption.weight(.semibold))
-                    .foregroundStyle(QuillCodePalette.muted)
-                TextField("quillcode-feature", text: $draft.path)
-                    .textFieldStyle(.roundedBorder)
-            }
-
-            VStack(alignment: .leading, spacing: 6) {
-                Text("New branch")
-                    .font(.caption.weight(.semibold))
-                    .foregroundStyle(QuillCodePalette.muted)
-                TextField("feature/quillcode", text: $draft.branch)
-                    .textFieldStyle(.roundedBorder)
-            }
-
-            VStack(alignment: .leading, spacing: 6) {
-                Text("Base ref")
-                    .font(.caption.weight(.semibold))
-                    .foregroundStyle(QuillCodePalette.muted)
-                TextField("main", text: $draft.base)
-                    .textFieldStyle(.roundedBorder)
-                Text("Leave branch or base blank to use git defaults.")
-                    .font(.caption)
-                    .foregroundStyle(QuillCodePalette.muted)
-            }
-
-            HStack {
-                Spacer()
-                Button("Cancel", action: onCancel)
-                Button("Create", action: onCreate)
-                    .buttonStyle(.borderedProminent)
-                    .disabled(!draft.canCreate)
-            }
-        }
-        .padding(24)
-        .frame(width: 520)
-        .background(QuillCodePalette.background)
-    }
-}
-
-private struct QuillCodeWorktreeRemoveView: View {
-    @Binding var draft: QuillCodeWorktreeRemoveDraft
-    var onCancel: () -> Void
-    var onRemove: () -> Void
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 18) {
-            HStack(alignment: .center) {
-                VStack(alignment: .leading, spacing: 4) {
-                    Text("Remove Worktree")
-                        .font(.title2.weight(.semibold))
-                    Text("Remove an existing registered git worktree.")
-                        .font(.callout)
-                        .foregroundStyle(QuillCodePalette.muted)
-                }
-                Spacer()
-                Image(systemName: "minus.rectangle")
-                    .font(.title2)
-                    .foregroundStyle(QuillCodePalette.yellow)
-            }
-
-            VStack(alignment: .leading, spacing: 6) {
-                Text("Worktree folder")
-                    .font(.caption.weight(.semibold))
-                    .foregroundStyle(QuillCodePalette.muted)
-                TextField("quillcode-feature", text: $draft.path)
-                    .textFieldStyle(.roundedBorder)
-                Text("Removal is limited to worktrees registered by git.")
-                    .font(.caption)
-                    .foregroundStyle(QuillCodePalette.muted)
-            }
-
-            Toggle("Force removal", isOn: $draft.force)
-
-            HStack {
-                Spacer()
-                Button("Cancel", action: onCancel)
-                Button("Remove", action: onRemove)
-                    .buttonStyle(.borderedProminent)
-                    .disabled(!draft.canRemove)
-            }
-        }
-        .padding(24)
-        .frame(width: 520)
-        .background(QuillCodePalette.background)
-    }
-}
-
 private struct QuillCodeTranscriptView: View {
     var transcript: TranscriptSurface
     var contextBanner: ContextBannerSurface?
@@ -1480,40 +827,6 @@ private struct QuillCodeContextBannerView: View {
         )
         .clipShape(RoundedRectangle(cornerRadius: 14, style: .continuous))
         .frame(maxWidth: .infinity, alignment: .leading)
-    }
-}
-
-private struct QuillCodeWorktreeCreateDraft: Equatable {
-    var path = ""
-    var branch = ""
-    var base = ""
-
-    var canCreate: Bool {
-        !path.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty
-    }
-
-    var request: WorkspaceWorktreeCreateRequest {
-        WorkspaceWorktreeCreateRequest(
-            path: path.trimmingCharacters(in: .whitespacesAndNewlines),
-            branch: branch.trimmingCharacters(in: .whitespacesAndNewlines),
-            base: base.trimmingCharacters(in: .whitespacesAndNewlines)
-        )
-    }
-}
-
-private struct QuillCodeWorktreeRemoveDraft: Equatable {
-    var path = ""
-    var force = false
-
-    var canRemove: Bool {
-        !path.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty
-    }
-
-    var request: WorkspaceWorktreeRemoveRequest {
-        WorkspaceWorktreeRemoveRequest(
-            path: path.trimmingCharacters(in: .whitespacesAndNewlines),
-            force: force
-        )
     }
 }
